requirements.txt
flask>=2.0.0
flask-socketio>=5.3.0
flask-cors>=3.0.10
requests>=2.28.0
docker>=6.0.0
argparse>=1.4.0
mysql-connector-python>=8.0.28
python-dotenv>=1.0.0
